+++ extensions/shared/library/src/main/java/app/morphe/extension/shared/Utils.java
@@ -417,6 +417,10 @@ public class Utils {
         clipboard.setPrimaryClip(clip);
     }
 
+    public static boolean isNotEmpty(@Nullable String str) {
+        return str != null && !str.isEmpty();
+    }
+
     public static boolean isTablet() {
         return context.getResources().getConfiguration().smallestScreenWidthDp >= 600;
     }