+++ extensions/shared/library/src/main/java/app/morphe/extension/shared/settings/preference/SharedPrefCategory.java
@@ -1,5 +1,6 @@
 package app.morphe.extension.shared.settings.preference;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.preference.PreferenceFragment;
@@ -37,17 +38,25 @@ public class SharedPrefCategory {
         removeKey(key);
     }
 
+    @SuppressLint("ApplySharedPref") // Must use commit to ensure default value is not saved to preferences.
     private void saveObjectAsString(@NonNull String key, @Nullable Object value) {
         preferences.edit().putString(key, (value == null ? null : value.toString())).commit();
+    }
+
+    @SuppressLint("ApplySharedPref") // Must use commit to ensure default value is not saved to preferences.
+    public void clear() {
+        preferences.edit().clear().commit();
     }
 
     /**
      * Removes any preference data type that has the specified key.
      */
+    @SuppressLint("ApplySharedPref") // Must use commit to ensure default value is not saved to preferences.
     public void removeKey(@NonNull String key) {
         preferences.edit().remove(Objects.requireNonNull(key)).commit();
     }
 
+    @SuppressLint("ApplySharedPref") // Must use commit to ensure default value is not saved to preferences.
     public void saveBoolean(@NonNull String key, boolean value) {
         preferences.edit().putBoolean(key, value).commit();
     }