+++ extensions/youtube/src/main/java/app/morphe/extension/youtube/settings/preference/SpoofStreamingDataSideEffectsPreference.java
@@ -92,20 +92,13 @@ public class SpoofStreamingDataSideEffectsPreference extends Preference {
             case ANDROID_VR_1_47_48, ANDROID_VR_1_54_20 ->
                     summary = str("morphe_spoof_video_streams_about_no_audio_tracks")
                             + '\n' + str("morphe_spoof_video_streams_about_no_stable_volume");
-            case ANDROID_NO_SDK ->
-                    summary = str("morphe_spoof_video_streams_about_playback_failure");
+            case TV, TV_SIMPLY ->
+                    summary = str("morphe_spoof_video_streams_about_js");
             case VISIONOS ->
                     summary = str("morphe_spoof_video_streams_about_experimental")
                             + '\n' + str("morphe_spoof_video_streams_about_no_audio_tracks")
                             + '\n' + str("morphe_spoof_video_streams_about_no_av1");
             default -> Logger.printException(() -> "Unknown client: " + clientType);
-        }
-
-        // Only Android No SDK can play children videos in incognito, but it commonly fails at 1 minute
-        // or doesn't start playback at all. List the side effect for other clients
-        // since they will fall over to Android No SDK.
-        if (clientType != ClientType.ANDROID_NO_SDK) {
-            summary += '\n' + str("morphe_spoof_video_streams_about_kids_videos");
         }
 
         // Use better formatting for bullet points.