+++ patches/src/main/kotlin/app/morphe/patches/shared/Fingerprints.kt
@@ -33,3 +33,9 @@ internal object BoldIconsFeatureFlagFingerprint : Fingerprint(
     )
 )
 
+internal object ProtobufClassParseByteArrayFingerprint : Fingerprint(
+    accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.STATIC),
+    returnType = "L",
+    parameters = listOf("L", "[B"),
+    custom = { method, _ -> method.name == "parseFrom" }
+)