+++ patches/src/main/kotlin/app/morphe/patches/youtube/layout/searchfilter/Fingerprints.kt
@@ -11,13 +11,6 @@ import app.morphe.patches.shared.misc.mapping.resourceLiteral
 import com.android.tools.smali.dexlib2.AccessFlags
 import com.android.tools.smali.dexlib2.Opcode
 
-internal object ProtobufClassParseByteArrayFingerprint : Fingerprint(
-    accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.STATIC),
-    returnType = "L",
-    parameters = listOf("L", "[B"),
-    custom = { method, _ -> method.name == "parseFrom" }
-)
-
 internal object SearchFilterDialogFingerprint : Fingerprint(
     accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.FINAL),
     returnType = "Landroid/view/View;",