+++ patches/src/main/kotlin/app/morphe/patches/youtube/misc/playservice/VersionCheckPatch.kt
@@ -66,6 +66,9 @@ var is_20_03_or_greater : Boolean by Delegates.notNull()
 var is_20_05_or_greater : Boolean by Delegates.notNull()
     private set
 @Deprecated("20.14.43 is the lowest supported version")
+var is_20_06_or_greater : Boolean by Delegates.notNull()
+    private set
+@Deprecated("20.14.43 is the lowest supported version")
 var is_20_07_or_greater : Boolean by Delegates.notNull()
     private set
 @Deprecated("20.14.43 is the lowest supported version")
@@ -121,7 +124,8 @@ var is_21_02_or_greater : Boolean by Delegates.notNull()
     private set
 var is_21_03_or_greater : Boolean by Delegates.notNull()
     private set
-
+var is_21_05_or_greater : Boolean by Delegates.notNull()
+    private set
 
 val versionCheckPatch = resourcePatch(
     description = "Uses the Play Store service version to find the major/minor version of the YouTube target app.",
@@ -151,6 +155,7 @@ val versionCheckPatch = resourcePatch(
         is_20_02_or_greater = 250299000 <= playStoreServicesVersion
         is_20_03_or_greater = 250405000 <= playStoreServicesVersion
         is_20_05_or_greater = 250605000 <= playStoreServicesVersion
+        is_20_06_or_greater = 250705000 <= playStoreServicesVersion
         is_20_07_or_greater = 250805000 <= playStoreServicesVersion
         is_20_09_or_greater = 251006000 <= playStoreServicesVersion
         is_20_10_or_greater = 251105000 <= playStoreServicesVersion
@@ -177,5 +182,6 @@ val versionCheckPatch = resourcePatch(
         is_20_49_or_greater = 255005000 <= playStoreServicesVersion
         is_21_02_or_greater = 260305000 <= playStoreServicesVersion
         is_21_03_or_greater = 260405000 <= playStoreServicesVersion
+        is_21_05_or_greater = 260605000 <= playStoreServicesVersion
     }
 }